refactor: Update CFImages namespace and add ServeImageParams and ServeImageResponse interfaces
This commit refactors the CFImages namespace in the index.ts file to include the new ServeImageParams and ServeImageResponse interfaces. These interfaces are used to define the parameters and response types for the new getImage method in the CFImages class. This change improves the code organization and makes it easier to understand and use the CFImages API.